+++ CIPMSWeb/Enrollment/Chicago/Step2_coupon.aspx.cs
@@ -56,7 +56,7 @@ public partial class Step2_Chicago_2_coupon : System.Web.UI.Page
 			if (isReadOnly)
 			{
 				DataSet dsApp = CamperAppl.getCamperApplication(strFJCID);
-				iStatus = Convert.ToInt16(dsApp.Tables[0].Rows[0]["Status"]);
+				iStatus = Convert.ToInt32(dsApp.Tables[0].Rows[0]["Status"]);
 			}
 			else
 			{
